#2 - Approve & apply leader's propose
